% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modefill.R
\name{modefill}
\alias{modefill}
\title{Fill missing DDD values}
\usage{
modefill(
  purch_data,
  actv_time = 60,
  pvmvar = "otpvm",
  vnrovar = "vnr",
  kustvar = "kust",
  dddvar = "ddd",
  dispatch_rm = TRUE,
  extra_rm = TRUE,
  rstr_ord = FALSE,
  print_info = FALSE,
  info_price = TRUE
)
}
\arguments{
\item{purch_data}{data frame of the drug purchase data}

\item{actv_time}{how long a value stays active to be able to predict from, 60 by default}

\item{pvmvar}{purchase date variable name in the data frame, "otpvm" by default}

\item{vnrovar}{drug code variable name in the data frame, "vnr" by default}

\item{kustvar}{cost variable name in the data frame, "kust" by default}

\item{dddvar}{ddd variable name in the data frame, "ddd" by default}

\item{dispatch_rm}{remove dispatch cost, TRUE by default}

\item{extra_rm}{remove all the added variables from the returned data, TRUE by default}

\item{rstr_ord}{restore original order of the data at the end of function, FALSE by default}

\item{print_info}{prints info of all the predictions made, FALSE by default}

\item{info_price}{show package cost in info instead of DDD / cost, FALSE by default}
}
\value{
data frame with filled DDD values
}
\description{
Fills the missing DDD values by finding comparative purchases (same vnr) from the data and calculating their DDD / cost. New DDD is then just this times the purchase cost.
Comparative purchases are found following logic: 1. search current interval where prices should have stayed the same (intervals are 1.-14. and 15.-30. of the month), and calculate the mode
DDD / cost. If no comparative purchases found, see for the latest active mode from previous intervals. How long a mode stays active depends on parameter actv_time (given in days). If
multiple values with themax frequency, calculate the median of these max frequency values.
}
